Larger refactoring to introduce additional BLE backends
This change is based on the discussion in https://github.com/open-homeautomation/miflora/pull/23
I moved everything related to gatttool into a separate module and class. I tried to change as little as possible to the structure and flow of the code.
The solution is *not* proberly tested. This is a first proposal to discuss the general ideas of the new architecture.

Add .idea to gitignore (used in PyCharm)